requirement.txt
asgiref==3.5.2
certifi==2022.9.24
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
protobuf==3.20.1
pytz==2022.6
sqlparse==0.4.3
